Rename `_config_managing.py` to `_managing_actor.py`
As this package gets more stuff, the old name used when the actor was
stored in the `frequenz.actor` package doesn't make much sense and is
confusing.
