+++ src/frequenz/sdk/config/__init__.py
@@ -3,8 +3,8 @@
 
 """Read and update config variables."""
 
-from ._config_managing import ConfigManagingActor
 from ._logging_config_updater import LoggerConfig, LoggingConfig, LoggingConfigUpdater
+from ._managing_actor import ConfigManagingActor
 from ._util import load_config
 
 __all__ = [
